+++ test/system/tests/activecredential.sh
@@ -53,7 +53,7 @@ echo "12345678" > secret.data
 
 tpm2_createek -Q -H 0x81010009 -g rsa -p ek.pub
 
-tpm2_getpubak -E 0x81010009 -k 0x8101000a -g rsa -D sha256 -s rsassa -f ak.pub -n ak.name > ak.out
+tpm2_createak -E 0x81010009 -k 0x8101000a -g rsa -D sha256 -s rsassa -p ak.pub -n ak.name > ak.out
 
 # Capture the yaml output and verify that its the same as the name output
 loaded_key_name_yaml=`python << pyscript