+++ test/system/tests/output_formats.sh
@@ -97,7 +97,7 @@ for fmt in tss pem der; do
 
 done
 
-tpm2_getpubak -Q -g $alg_ak -E $handle_ek -k $handle_ak -f "$file_pubak_tss" -n "$file_pubak_name"
+tpm2_createak -Q -g $alg_ak -E $handle_ek -k $handle_ak -p "$file_pubak_tss" -n "$file_pubak_name"
 
 tpm2_readpublic -Q -H $handle_ak -f "pem" -o "$file_pubak_pem"