+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Changelog
 ### next
+  * tpm2_getpubak renamed to tpm2_createak, -f becomes -p and -f is used for format of public key
+    output.
   * tpm2_getpubek renamed to tpm2_createek, -f becomes -p and -f is used for format of public key
     output.
   * Libre SSL builds fixed.